feat: add shopping item api

+++ test/Groceries.Boudreau.Cloud.Integration/TestShoppingListContextFixture.cs
@@ -11,6 +11,8 @@
     /// </summary>
     public class TestShoppingListContextFixture : IDisposable
     {
+        private static bool calledOnce = false;
+
         public IConfigurationRoot Configuration;
 
         public string ConnectionString
@@ -30,19 +32,27 @@
                .AddJsonFile("appsettings.json", optional: true, reloadOnChange: true)
                .AddEnvironmentVariables()
                .Build();
-            
+
             var dbContextOptionsBuilder = new DbContextOptionsBuilder();
             dbContextOptionsBuilder.UseSqlServer(ConnectionString);
             DbContextOptions = dbContextOptionsBuilder.Options;
 
-            using(var context = new ShoppingListContext(DbContextOptions))
+            if (calledOnce)
+            {
+                return;
+            }
+
+            using (var context = new ShoppingListContext(DbContextOptions))
             {
                 if (Configuration["SkipDeleteDatabase"] == null)
-                { 
+                {
                     context.Database.EnsureDeleted();
                 }
+
                 context.Database.Migrate();
             }
+
+            calledOnce = true;
         }
 
         public void Dispose()
